Adds indicator signal for main
Adds ### indicator to point the location of the main function. Notice that this function is different than the handler. the main function performs all the operations and requests of the pipeline. The handler function is used by GCP to trigger main once the Cloud Function endpoint is requested with a POST method.